drbd: fix splitting of large discard requests
This is a fixup to the recent commit 06bbd6eec1
'drbd: remove plausible_request_size()'
The bio size needs to be checked __after__ blk_queue_split().

The effect of the bug was that a 'blkdiscard' command failed on DRBD devices
(when the device is larger than the discard granularity). With that change
discards on the whole device were not properly split into the necessary pieces.