drbd: do not merge discards after limiting them for alignment
This is a fixup for:
751b90de9381 drbd: enable rs-discards if backend discard_granularity > rs_discard_granularity

That commit caused the flags EE_RS_TRIM_LIMITED_FRONT and
EE_RS_TRIM_LIMITED_BEHIND to be added to the peer request to enforce
alignment. However, the request could still be merged across these
limits, leading to discards being submitted that were too large. This
caused zeros to be written unnecessarily.

Prevent this by checking the limit flags when deciding whether to merge
discards.