drbd: also propagate UUIDs when we are WFBitMapS to a peer
The idea from
8dee85826642 drbd: Propagate UUIDs when I am source of a resync operation
also applies to peers with replication state WFBitMapS. Add this missing
case.

One issue that this missing case caused was an unnecessary full resync.
At the end of the current resync, the target set its current UUID to
what was the source's current UUID at the start of the resync. So it was
UpToDate, but its current UUID was an older one. When a later event
caused the nodes to compare UUIDs again, the presence of the older UUID
caused a full resync to be started.