drbd: do not mask extra bits when copying bitmap slot
This code was buggy and unnecessary. Remove it.

The bug was detected as a "UBSAN: shift-out-of-bounds" warning for
drbd_bm_copy_slot(). This occurs for specific combinations of backing
device size and bm_max_peers value. "bitmap->bm_words" is the number of
64 bit words that cover the used area of the bitmap. Unfortunately it
can overestimate.

To illustrate, imagine we allowed a device with 1 bit of bitmap per peer
and bm_max_peers == 2. Then on a 64 bit system, we get
bm_words == (ALIGN(bits, 64) * b->bm_max_peers) / BITS_PER_LONG == 2

Thus the layout of the bitmap in 32 bit words is:
[peer 0][peer 1]  # should have been end of bitmap
[unused][unused]  # end of bitmap according to bm_words

As a result, the final iteration in drbd_bm_copy_slot() ends up in an
area of the bitmap that is entirely unused. The bit masking code failed
to take this case into account, causing a negative shift to be
calculated.

The code was unnecessary anyway. These bits should not be set since they
are beyond the used area of the bitmap. If something does set them, then
we should fix that there. There is no need to patch up the incorrectly
set bits here.

It would seem sensible to fix the definition of bm_words. However that
would cause problems exchanging bitmaps with older peers, so it is
non-trivial.

Co-developed-by: Lars Ellenberg <lars.ellenberg@linbit.com>