drbd: remove lru_cache, use upstream version
Turns out this is pretty much the same as what's in the Linux kernel
already (we upstreamed it ages ago). Our version did get a few fixes over
time, but it does not seem like it's any more or less broken than the
upstream version.

Unfortunately, we cannot completely remove lru_cache, because some
distro kernels do not enable CONFIG_LRU_CACHE.

So move the compat implementation to drbd-kernel-compat and only use it
if lru_cache is not available upstream.