drbd: do not call dec_rs_pending() when clearing up "trims"
If the flag EE_TRIM is set on a resync request, then the reply for this
request has already been received. Specifically,
receive_rs_deallocated() was called. The request remains on the list
resync_request_ee until any potential discard merging is complete.

We have already called dec_rs_pending() for such a request, so we must
not call it again when clearing up on connection loss.

This is a "quick fix" for the upcoming release. It may be better to fix
this by changing which list such requests are placed on.