compat: test and patch for linux/sched/signal.h
Turns out I removed a little too much "compat cruft" in commit
99d5d254ae06 ("compat: remove ancient include/linux compat").

Some kernels still don't have linux/sched/signal.h. So, bring the compat
back, but do it with a cocci patch this time.

In case this header does not exist, simply remove the include. This
restores the behavior from before where we just included an empty header
in the compat case.