drbd: count nodes in the quorumless_nodes set as outdated for calculating quorum
strictly speaking they are not D_OUTDATED, but maybe quorum lost primary
(I/O frozen or I/O throwing errors, etc...), or forced to secondary, but
for sure without quorum. For our consideration, it is good enough to
let them not be added to the "unknown" category.

Explaining the benefit of all this by example. Consider

    C       | C is diskless
   / \      | A is Primary and has a backing disk
  A---B     | B is Secondary, having a backing disk

When A crashes...

    C       | C is diskless
     \      | A is dead
      B     | B has quorum with the help of tie-breaker C

When C is removed/downed

            | C is down
            | A is dead
      B     | B keeps quorum since A is part of the quorumless_nodes mask

That B keeps quorum in this situation is new and desired!
The other critical case is:

    C       | C is diskless
   / \      | A is Secondary, having a backing disk and quorum
  A---B     | B is Secondary, having a backing disk and quorum

The connection A-B is interrupted

    C       | C is diskless
   / \      | A is Secondary, having a backing disk and quorum
  A   B     | B is Secondary, having a backing disk and quorum

C goes down

            |
            | A is Secondary no quorum
  A   B     | B is Secondary no quorum

Previous attempts failed in this scenario. The "keep tie-breaker quorum"
idea granted quorum to A and B in this scenario. Which is fatal, of course.