drbd: send ping acks using unbound workqueue
The standard system workqueue can exhibit scheduling latency spikes.
This causes the ping ack work function to be executed too late, and the
peer drops the connection with the message:
drbd res peer: PingAck did not arrive in time.

This has been observed specifically on a CentOS 7 system. Latencies
spiked every 5 minutes. In a 24 hour period, 100 cases of latency over
5s were recorded. These spikes coincide with 100% CPU consumption from
the real-time process "migration":
  PID USER      PR  NI    VIRT    RES    SHR S  %CPU %MEM     TIME+ COMMAND
   85 root      rt   0       0      0      0 S 100.0  0.0  56:09.52 migration/15

Sending ping acks is a latency-sensitive operation, so use an unbound
high priority workqueue to send them instead of the standard workqueue.
We also set WQ_MEM_RECLAIM because DRBD may be used in the memory
reclaim path. When under memory pressure we want to keep the DRBD
connection alive so that we can write data out over it.