compat: test and patch for holder in fs_dax_get_by_bdev
Since upstream commit 8012b8660855 ("dax: introduce holder for dax_device"),
fs_dax_get_by_bdev takes a third and fourth parameter, used to identify the
holder and register some callbacks in case of errors.

This commit adds a new compat test that checks for the 4 parameter variant
and falls back to the older patch for 2 and 1 parameter variants.