compat: update fill-in for fs_dax_get_by_dev
In a68200c121de we updated the call to fs_dax_get_by_bdev to take 4 arguments,
in line with the currrent kernel. But we forgot to update the replacement
should fs_dax_get_by_bdev not be available on the kernel. The patch
needs to be updated to also look for the 4 arguments.