drbd: remove PARANOIA
The only user of the PARANOIA define is the IS_VALID_MDEV machinery. The
last user of that was removed in commit cbc66a141b24 ("r2394: This huge
patch took me the better part of the last couple of weeks [...]") in
2006(!)

Remove the unused IS_VALID_MDEV macro and the then-unused PARANOIA
define.