drbd: control all CONFIG_ options via Kbuild
It is rather confusing to have a CONFIG_ option overridden in some
header. Instead, do this in Kbuild where we can append the option to the
"global" CFLAGS.

This keeps the current behavior of enabling fault injection by default
for the out-of-tree module, even though I'm not convinced that is
actually desirable.