drbd: remove macros using require_context
This require_context attribute originated in a proposed sparse patch by
Philipp Reisner back in 2008. Johannes Berg had a different solution to
a similar problem, and that patch "won" in the end; so the require_context
thing never got merged. The whole history can be read at [0].

DRBD kept using these annotations anyway for a while. Nowadays, on a
modern unmodified sparse, they obviously do nothing, and they are hardly
used anymore anyway.

The exception is __must_hold, but that is already defined in
<linux/compiler_types.h>, so do not redefine it. See also Linux kernel
commit d05d199883b0 ("drbd: Do not redefine __must_hold()").

So, just remove the definitions of these macros.

It also seems like the ctags ignore list in the Makefile has not been
checked in a long time, since none of these identifiers are actually
defined anymore.

[0] https://www.spinics.net/lists/linux-sparse/msg01150.html