drbd: propagate reachable_nodes mask with P_TWOPC_COMMIT
The quorumless_nodes mask is beneficial for keeping quorum when an
already degraded cluster's configuration is changed. Commit
2ae49562374 introduced this concept. A few days later, with commit
87623754e15, I had to remove triggering additional two-phase-commits
for leaving secondaries.

So far, only the initiator of a two-phase-commit became aware of all
of the cluster's member nodes at commit time. From now on, all nodes
get that valuable information by sending the reachable_nodes mask with
the P_TWOPC_COMMIT packet.
The recipients can then use that to update their quorumless_nodes
mask.

This change compensates for the disadvantage of commit 87623754e15
without increasing the number of two-phase commits.

With this patch, after a crash of a primary, all remaining nodes have
the crashed node masked in their quorumless_nodes mask.
(Before this patch, only the neighbors of that crashed primary had it)

With this patch, after a graceful disconnect of a secondary, all
remaining nodes have the disconnected node masked in their quorumless_nodes
mask. (Before this patch, only the initiator of the disconnect had it)

Explaining the change's benefit by example. Consider nodes A, B, and C
with ascending node IDs.

        C       | C is diskless
       / \      | A is Secondary and has a backing disk
      A---B     | B is Secondary, having a backing disk

Step 1: A does a graceful disconnect from B

        C       | C is diskless and has quorum
       / \      | A is Secondary with disk and has quorum
      A   B     | B is Secondary with disk and has quorum

Step 2: A does a graceful disconnect from C

        C       | C is with quorum since it is part of 2 out of 3
         \      | A is **without** quorum
      A   B     | B is with quorum since it is part of 2 out of 3

Step 3: C does a graceful disconnect from B

        C       | C is without quorum
                | A is without quorum
      A   B     | B is ...

Before this change, B has lost quorum as well. With this change, B keeps
quorum, because it knows that A is without quorum. This patch enables
that, in step 2, node B learns during the two-phase-commit that A is now
an isolated 1 out of 3. Hence it can not have quorum. It uses this
information in step 3 and keeps quorum.

Crashing secondaries still need to be added to the quorumless_nodes
mask by future improvements.