drbd: Allow state transition D_NEGOTIATING -> D_OUTDATED
It is valid that the disk state can go from D_NEGOTIATING to
D_OUTDATED.  This may happen when two nodes are outdated, and one of
them does detach and attach.

A pair of outdated nodes are uncommon. A way to get there is in a
four-node cluster. Two nodes leave gracefully, one by one. One of the
remaining cluster nodes is primary, which causes the leaving nodes to
outdate themselves. Then they connect and form a quorum-less partition
of two outdated nodes.

This patch fixes a detach/attach operation in that outdated partition.