drbd: allow got_BlockAck() to safely find resync requests
This fixes a race condition which resulted in connection loss with
messages of the form:
drbd res/0 drbd1 peer: Began resync as SyncSource (will sync 307144 KB [76786 bits set]).
drbd res/0 drbd1 peer: Unexpected resync write ack at 60664s+4096
drbd res peer: conn( Connected -> NetworkFailure ) peer( Secondary -> Unknown )

The issue affected sync source nodes. The problem was that
w_e_end_rsdata_req() added the resync request to the list resync_ack_ee
after sending the reply. So the ack could be received before the request
was added to the list. In that case the request could not be found and
the connection was terminated.

This issue has been present since the beginning of DRBD 9.2,
specifically:
223009b6369f drbd: synchronize resync with interval tree and dagtags

Add the request to resync_ack_ee before sending the reply. We must avoid
freeing the request in got_BlockAck() while we still have a reference to
it in drbd_rs_reply(). We achieve this by freeing the request only when
both the "sent" and "received" flags have been set.

Note that it is still safe to access the requests via a list traversal
protected by peer_reqs_lock, as reclaim_from_resync_ack_ee() does. The
presence of the request on the list acts as an additional reference.

Co-developed-by: Philipp Reisner <philipp.reisner@linbit.com>