drbd: Fix a quorumless_nodes aspect for four nodes or more
Commit d25fc604c80 introduced the idea of counting quorumless nodes
the same way as outdated nodes. But that is flawed.

"outdated" means the other node promises not to promote because it
 knows that it lacks the latest version of the data.

"quorumless" means the other node was guaranteed not to have a quorum
 at some point. When more nodes join it, and the partition of
 quorumless nodes becomes the majority, they are no longer quorumless.
 Another way of expressing it is: the quorumless node mask is a
 way to keep quorum when the diskless tiebreaker goes away.

Here is an example that highlights what was flawed so far.
A, B, C, and D are fully connected. A is primary.

1. A gets isolated and loses quorum.

2. B, C, and D keep quorum as the majority. As a consequence, they
   mark A as quorumless.

3. C gets promoted, and B disconnects gracefully. With that, B outdates
   itself.

4. C and D keep quorum as they are 2 out of 2. A is quorumless, and B is
   outdated.

5. C demotes, and D promotes. C disconnects gracefully and outdates
   itself.

D keeps quorum as it is 1 out of 1. A is quorumless, and B and C are
outdated.

Here is the pitfall. The problem is that A, B, and C are 3 of 4. They
get quorum, and A has an UpToDate disk, and A, B, and C can become a
partition with quourm and UpToDate data.

It needs to reject the graceful disconnect in step 5. This happens
when the quorum implementation recognizes at that point that A, B, and
C can gain quorum together. Therefore the C and D partition would lose
quorum. Therefore it rejects the graceful disconnect so that the primary
D keeps quorum.

This is what this change improves.