ci: also build branches in downstream pipelines
We already rebuild DRBD in a downstream pipeline. However, this relies
on the "$CI_COMMIT_BRANCH == $RELEASE_BRANCH" rule. We want to be able
to build branches in downstream pipelines too. Add rules to enable this.

Also add the option to make the build conditional. Even when nothing is
built, the pipeline is required to determine DRBD_VERSION.