drbd: remove unnecessary flags for drbd_find_conflict()
Previously, drbd_find_conflict() tried to avoid testing the type of the
given interval. This is an unnecessary abstraction. Only keep the flags
that depend on the calling context and hence cannot be determined from
the given interval.

This makes an upcoming change to prevent sending conflicting resync
requests simpler.