drbd: introduce struct twopc_request
Up to now, the code has re-used p_twopc_request from the protocol
definition. I am replacing this now with an in-core representation
struct twopc_request. The advantage is that we can support different
packet versions when encoding/decoding. This change is also a
preparation step for the next commit.