drbd: remember online nodes instead of quorumless nodes
It is a lot easier to think about online nodes than about nodes which
are guaranteed to have lost quorum. (and less code!)

We cache a node mask of the online members of the cluster. It might
be off because a node is still marked as online immediately after
it crashes. That means it might have an online mark for an already
offline node. On the other hand, we guarantee that it never has
a zero for an online node.

In the places where we need the quorumless nodes, we can use the
inverse bitmask of members if
  * our parition has quorum
  * and the other partition can not have quorum.

Using this principle in __calc_quorum_with_disk() and
__calc_quorum_node_disk(), and the check is right behind the calls
in calc_quorum().