drbd: initiate an empty two-phase-commit after a secondary peer crashed
Here is the second attempt to do this. The first time I had to remove
it (see 87623754e1) because it caused too many two-phase-commits.

Since 27aea8f9cc we get resource->members updates with every 2PC, so
we trigger this additional two-phase-commit only for unexpectedly
leaving secondaries.

Explaining the change's benefit by example. Consider nodes A, B, and C
with ascending node IDs.

        C       | C is diskless
       / \      | A is Secondary and has a backing disk
      A---B     | B is Secondary, having a backing disk

Step 1: A does a graceful disconnect from B

        C       | C is diskless and has quorum
       / \      | A is Secondary with disk and has quorum
      A   B     | B is Secondary with disk and has quorum

Step 2: A disconnects unexpectedly from C (with --force or crash)

        C       | C is with quorum since it is part of 2 out of 3
         \      | A is **without** quorum
      A   B     | B is with quorum since it is part of 2 out of 3

Step 3: C disconnects unexpectedly from B (with --force or crash)

        C       | C is without quorum
                | A is without quorum
      A   B     | B is ...

Before this change, B has lost quorum as well. With this change, B keeps
quorum, because it knows that A is without quorum. This patch enables
that, in step 2, node B learns during the two-phase-commit that A is now
an isolated 1 out of 3. Hence it can not have quorum. It uses this
information in step 3 and keeps quorum.