drbd: do not set bitmap bits when attaching to an up-to-date disk
This is relevant when attaching to a disk when we were previously a
diskless client. In particular, when the day0 UUID has been set and the
disk marked as up-to-date. That is usually used to mean that the disk is
all zeros or is thinly provisioned.

Note that in this context, "up-to-date" does not mean that the data is
necessarily the newest in the whole cluster. It just means that it is
up-to-date with the current UUID stored in the metadata.