drbd: fix race freeing peer request in drbd_resync_request_complete()
This race was observed when running a checksum based resync where some
of the intervals were in sync and some were not. In this case, some
requests are freed by the ack_sender in e_end_resync_block(), while
others are freed in ack receiving context in got_IsInSync(). As a
result, drbd_resync_request_complete() may be called concurrently for
different requests. After one thread sets INTERVAL_COMPLETED, the other
might then find the request in drbd_check_peers_in_sync_progress() and
free it. Hence we must not use the peer request after setting
INTERVAL_COMPLETED.

The errors logged as a result of this bug were varied. Here is one
example:

[  130.647521] drbd res/0 drbd1 peer: Began resync as SyncTarget (will sync 26240 KB [6560 bits set]).
[  131.335012] BUG: unable to handle kernel NULL pointer dereference at 0000000000000090
[  131.335899] IP: [<ffffffffa29ccd87>] put_page+0x17/0x60
...
[  131.335899] Call Trace:
[  131.335899]  [<ffffffffc05b0760>] drbd_free_pages+0x150/0x1b0 [drbd]
[  131.335899]  [<ffffffffc05ba772>] __drbd_free_peer_req+0x52/0x130 [drbd]
[  131.335899]  [<ffffffffc05bdc8a>] drbd_resync_request_complete+0x15a/0x1d0 [drbd]
[  131.335899]  [<ffffffffc05bde0e>] got_IsInSync+0x10e/0x390 [drbd]