drbd: cancel thin resync requests while source is starting resync
This fixes two cases that both led to this message being printed 10
times per second until an ongoing resync completed:
postponing start_resync ...

In both cases, the issue occurred when the sync target started sending
resync requests before the sync source had managed to execute
drbd_start_resync() and transition into SyncSource state.

In one case, a thin resync was configured.

In the other case, the resync was started with "invalidate-remote".

In these cases, the sync source responded to the resync requests even
though drbd_start_resync() had not yet run. This caused rs_pending_cnt
to be positive, and so do_start_resync() postponed the resync start. If
the resync was sufficiently active, then this repeated until the target
finished the resync.

After this was over, the sync source node transitioned into SyncSource
state and got stuck there, because it had missed the end of the resync.