drbd: Avoid a two-phase-commit retry in a specific situation
Consider a cluster of nodes A, B, and C with ascending node IDs.
B and C are connected:

     B
      \
   A   C

Then A initiates the two-phase commit to establish the connection A-B.
In the process, B forwards the P_TWOPC_PREPARE and P_TWOPC_COMMIT
messages to node C.

A considers the two-phase commit as completed in the moment it sent
the P_TWOPC_COMMIT packet towards B.

     B
    / \
   A   C

A moment later, A initiates the two-phase commit to establish the
connection A-C. The P_TWOPC_PREPARE of the second two-phase commit may
arrive at C before the P_TWOPC_COMMIT of the first two-phase commit.

Up to now, C then aborted the second transaction, and A needed to
retry it.

With this change, C will instead delay processing the P_TWOPC_PREPARE
until the first transaction is finished. It can only do that when the
two transactions have different "parents". That means when the packets
come over different connections.

  If we would wait for a P_TWOPC_COMMIT on the same connection where
  we got the P_TWOPC_PREOARE, we would deadlock the machinery.

     B
    / \
   A---C