drbd: allow peer to promote with the P_STATE packet
Normally, peer promotion occurs when the P_TWOPC_COMMIT packet is
received at the end of the two-phase commit. However, there are race
conditions that can cause P_STATE to be sent with the updated state
before P_TWOPC_COMMIT. Prior to this commit, such a race this could
cause the initial resync to be skipped.

Specifically this problem occurred because receive_state() only applied
the change of peer disk state to D_UP_TO_DATE, not the peer role change
to R_PRIMARY. Then when the P_TWOPC_COMMIT was received, only the peer
role was changed. As a result, neither state change fulfilled the
condition in finish_state_change() to set CONSIDER_RESYNC.

Such a race condition can occur when "resync-after" is used. The other
resource starting a resync causes our "resync-susp" status flags to be
changed. This change can occur while we are executing a two-phase
commit.

Fix this by allowing the peer role to be changed by receive_state(). We
trust our peer. If it claims to be Primary, then it must be in the
process of completing the promotion two-phase commit.

The exception removed by this commit was introduced by
cf32d08baf6e drbd: Update the peer role with the reachability packets

This commit was actually concerned with demotion rather than promotion.

As it happens, peer demotion with the P_STATE packet was then enabled in
1582135b67cc drbd: Use the P_STATE packet to inform peers about demotion

But promotion remained excluded. There never was a reason to disallow
peer promotion with P_STATE, so it is OK to enable it.