drbd: fix drbd_adm_detach()
Commit 5c02c258c2 replaced some enums with an unsigned int. That needs
to be a signed int. drbd_adm_finish() expects a signed int.

The problem became apparent when an adm_detach() freed the ldev
even when the state change failed. Because the line:

	if (retcode >= SS_SUCCESS) {

evaluates to true, when retcode holds a negative value and it is
declared to be an unsigned int.