drbd: do not trigger twopc after graceful disconnect of peer
The commit
9e4e702b580f drbd: initiate an empty two-phase-commit after a secondary peer crashed
introduced this code. The commit message makes it clear that it was only
intended to cover connection loss cases, rather than graceful
disconnects. A two-phase commit after a graceful disconnect is not
necessary because the initiator of the disconnect sends us the
appropriate new reachable_nodes.

This change would also prevent the problematic sequence described in the
recent commit "drbd: allow state change to continue when twopc pending
after lost peer", if that commit did not already do so.