drbd: Improve setting up request queue limits on diskless nodes
So far, we have failed to do anything useful with the queue limits on
diskless nodes. Moving forward, we will have physical and logical
block sizes right.

Well, as long as we connect to the diskfull node with the largest
physical sector size before promoting to primary.

So far, we called blk_set_stacking_limits() followed by a
blk_stack_limits() while establishing a new connection. That opened a
short window in which the block subsystem might create too large BIOs.