drbd: resume resync-after dependency chains that are not in order
This fixes a bug that occurred as follows. Suppose we have minors 1, 2
and 3, with:
* 2 should resync after 3
* 1 should resync after 2
* 3 is currently in a resync
* 2 does not need to perform a resync
* 1 does need to perform a resync

Now 1 and 2 have the resync suspended bit set while 3 is performing its
resync. When 3 completes its resync, 2 is resumed, but this has no
further effects because 2 does not need to perform a resync. Now, if no
other events occur to resume the resync of 1, it gets stuck in the
paused state.

Fix this by calling drbd_resume_next() until the state settles.

Also apply the same fix to the code path when pausing.