drbd: allow state change to continue when twopc pending after lost peer
The commit
9e4e702b580f drbd: initiate an empty two-phase-commit after a secondary peer crashed
caused the flag TWOPC_AFTER_LOST_PEER_PENDING to be set when a peer is
lost or disconnects.

The commit
3e7fefb34a86 drbd: delay promotion until try_become_up_to_date() has run
caused SS_NO_UP_TO_DATE_DISK to be returned from the state change
attempt when TWOPC_AFTER_LOST_PEER_PENDING was set.

The commit
de32acd7ad1f drbd: Only allow local promotion if fencing of all peers was successful
caused CS_FP_LOCAL_UP_TO_DATE to be set when the state change return
value is SS_NO_UP_TO_DATE_DISK and we have no peers to fence.

This combination of commits made the following sequence with nodes A, B
and C possible:
* A is primary
* An attempt is started to promote C. This waits in the "rv ==
  SS_TWO_PRIMARIES && !retried_ss_two_primaries" case of
  drbd_set_role().
* A is demoted and gracefully disconnected from C and B
* C sets TWOPC_AFTER_LOST_PEER_PENDING due to the disconnect of A, but
  does not run twopc_after_lost_peer() because state_sem is held by the
  promoting thread.
* C continues in drbd_set_role() and retries the promotion state change.
* This state change fails with SS_NO_UP_TO_DATE_DISK due to the
  TWOPC_AFTER_LOST_PEER_PENDING flag.
* The fencing case triggers, even though there is no fencing occurring
  here, causing CS_FP_LOCAL_UP_TO_DATE to be set.
* C promotes successfully, and generates a new UUID to due
  CS_FP_LOCAL_UP_TO_DATE, which is sent to B without
  UUID_FLAG_NEW_DATAGEN.
* B receives this new UUID, but due to the missing flag does not update
  its own current UUID.
* B and C are now both UpToDate and have different current UUIDs, but
  they do not determine any reason to trigger a resync.
* This causes problems later, such as spurious "split-brain" between A
  and B.

This code was added as an additional precaution. It is not needed to fix
the normal case for which commit 3e7fefb34a86 was introduced. The check
in after_primary_lost_events_settled() ensures that
try_become_up_to_date() runs before the attempted state change for the
promotion. Remove the code to prevent the sequence described above.

Fixes: 3e7fefb34a86 ("drbd: delay promotion until try_become_up_to_date() has run")