drbd: simplify resync discard merging
Replace the implementation of merging P_RS_DEALLOCATED packets with a
simpler one.

Previously, the drbd_peer_request structs were merged and freed as soon
as 2 adjacent P_RS_DEALLOCATED replies were received.

Now, we keep the drbd_peer_request structs and track the start and end
of the oldest resync discard range using pointers "received_last" and
"discard_last".

There are some changes in behavior, none of which are very significant:
* Discards will always be submitted in the order we made the requests.
  The previous implementation could submit a later discard before an
  earlier one when the deallocated reply and the replies for the
  surrounding requests had all been received while the earlier discard
  was pending.
* We consume a small amount of memory for longer for tracking the merged
  requests until the discard is complete.
* Discards that are waiting to be merged are no longer immediately
  submitted after sending the last resync request. This caused oddly
  sized discards to be submitted.

The main justification for this change is the reduction in code
complexity. The change also prepares us for sending an ack for every
resync reply. This may be necessary to allow the acks to be matched to
the resync replies more reliably.