drbd: use block_id to identify resync requests instead of ID_SYNCER
This makes the matching of replies and acks to the corresponding
requests more robust.

This fixes a bug which led to connection loss with a message like the
following on the SyncSource:
Unexpected resync write ack at 49024s+65536

With nodes P (diskless Primary), S (SyncSource) and T (SyncTarget), the
issue occurred as follows:

* T sends resync request R1 for interval I
* S reads I and replies to R1
* T writes the data; in e_end_resync_block() it removes R1 from the
  interval tree
* T moves resync_next_bit back for some reason (perhaps P_OUT_OF_SYNC
  due to writes from P from before the resync fully started)
* T sends resync request R2 for interval the same interval I
* T sends resync write ack for R1
* S receives R2
* S receives resync write ack for R1; in got_BlockAck() it finds and
  potentially frees the peer requests for both R1 and R2
* The above message occurs a bit later when S receives the resync write
  ack for R2

With precise matching of resync requests to resync replies this cannot
happen.

Online verify requests on the VerifyT side are now left in the interval
tree until the corresponding result packet has been received. This is
necessary so that we can match the result to the original request.