drbd: remove pointless wait
There may have been a legit reason for this delay in the past. Today
it makes no sense to delay the two-phase-commit after establishing the
TCP connections.

This caused a 100ms delay when thre is one peer. If the resource has 7
connections, it caused 700ms delay! This did nothing good.