drbd: do not allow transition from paused sync to bitmap exchange
As described in the recent commit "drbd: use mod_timer() instead of
add_timer() for starting resync", we may receive UUIDs from the peer
while we are in PausedSyncS state. We do not want to switch to WFBitMapS
in this case. The peer may still be in WFBitMapT state and does not
expect to receive the bitmap again.