drbd: align timeouts
Since the commit 6caee0f43b DRBD waits if the expected packets are
coming even if a two-phase-commit gets aborted. So far, that timeout
was ping-timeout * 4.

On the other hand, we used twopc_timeout / 8 as timeout to wait for
local two-pc aborts.

Connect retry breaks when a customer chooses a ping-timeout of 4
seconds and a two-pc-timeout of 10 seconds.

In wait_initial_states_received(), we wait for the remote to get the
P_TWOPC_PREPARE packet, and it replies with sizes, uuids, and state
packets. So the ping-timeout. It is two times the network latency.

The timeout in abort_local_transaction() should be a bit larget than
the timeout in wait_initial_states_received(), so let's use
ping-timeout * 1.5 there.