drbd: simplify peer request lists
The old concept was to keep the peer requests on some list with w.list
the whole time. Or at least most of the time. This led to some
situations were it was not clear which list was most appropriate. Adding
new lists for every situation is impractical. Mistakes in this area
caused DRBD to get stuck disconnecting.

It turns out these lists were largely unnecessary. Instead, we put the
peer requests on one of a couple of lists using the recv_order field,
and leave them there until we are finished with them. We still need
w.list, but only for specific situations.