drbd: Fix the in_handshake heuristic in is_valid_soft_transition()
When DRBD executes a handshake, it declines role changes (promote and
demote), attach, and detach operations. They would interfere with
establishing the connection and should be delayed until the connection
up.

The INITIAL_STATE_SENT bit has been used for this purpose so far. The
idea was that it gets set at the beginning of the handshake. As soon
as the repl_state[NOW] is no longer L_OFF the handshake must be over.
That is all correct.

The hidden bug is that it also gets set when a device gets added to
one node, and not on the other nodes.  As a consequence, the mentioned
detach, attach, promote, and demote operations look blocked on that
node.

The way out is to look at the HOLDING_UUID_READ_LOCK, which is only
used by the two-phase commit to establish a connection.