drbd: cancel requests waiting for conflicts on disconnect
Commit
f6acbd375397 drbd: ensure that we do not send conflicting resync requests
made a deadlock possible when disconnecting during a resync.

This occurred on a SyncTarget node as follows:
* Resync peer request R1 created and request sent to peer
* resync_next_bit jumps back, perhaps due to P_OUT_OF_SYNC
* Resync peer request R2 created, overlapping with R1; R2 held due to
  conflict
* Peer responds with data for R1, peer request submitted
* Disconnection

Now R1 progresses to done_ee, but is waiting for drbd_finish_peer_reqs()
to be called before being removed from the conflict tree. Meanwhile, R2
is on sync_ee and is held in the conflict tree. drain_resync_activity()
blocks waiting for sync_ee to drain.

Fix this by clearing resync requests that are waiting for conflicts
before waiting for sync_ee to drain.