compat: fixup req_write__yes_present patch
When we apply this patch, that means we are using the legacy REQ_* flags
(as opposed to REQ_OP_*).
But we still do not know whether or not we have enum req_op. The
enum_req_op__no_present patch may or may not be applied, so handle both
cases by rewriting all enum req_ops to unsigned ints; adjust the
following rules accordingly.