ci: de-duplicate the list of all kernels
A recent lbbuild commit added some extra build specs for sles15 builds
with "LB_NOSIGN" enabled. In our case, this would then end up as a
duplicate "<image> <kernel>" entry.

Apparently, lbbuild does not like it when multiple builds with the same
kernel version are started at the same time. It would be a waste of time
anyway...

So run the list of images/kernels through jq's "unique".