drbd: generalize diskless_with_peers_different_current_uuids()
Consider nodes A, B, and C, where B is intentionally diskless:

   B     | with B diskless and primary. It writes some data.
  / \    | B demotes to secondary.
 A---C   | A and C have backing disks.

   B     | C becomes isolated. B promotes, writes and demotes
  /      | again.
 A   C   | C is now an older data generation.

   B     | B and C establish a connection.
  / \    |
 A   C   |

When a diskless node connects to a node that pretends having UpToDate
data and is on another current-uuid the diskless node needs to figure
out if its exposed-uuid is newer or older than the data of the new
UpToDate peer.
So far, the code has covered the case when the diskless node is
primary. Starting with this change, it also considers that the
diskless node might be secondary. It either adopts the joining node's
current-uuid or outdates the joining node if it has older data.

The according code in santize_state() needs to follow the update. Note
that in both places, it triggered if the diskless node was
primary. This change removes that. Instead, it now checks if the
exposed_data_uuid already has a value.